server/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.2
tortoise-orm==0.25.1
aiomysql==0.2.0
aiosqlite==0.21.0
redis==6.2.0
aioredis==2.0.1
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydantic-extra-types==2.10.5
python-jose==3.5.0
passlib[bcrypt]==1.7.4
cryptography==45.0.5
httpx==0.28.1
python-multipart==0.0.20
pillow==11.3.0
aiosmtplib==4.0.1
Jinja2==3.1.6
python-dotenv==1.1.1
PyYAML==6.0.2
loguru==0.7.3
geoip2==5.1.0
maxminddb==2.8.2
python-dateutil==2.9.0
PyMySQL==1.1.1
rsa==4.9.1
websockets==15.0.1
psutil==7.1.3
aiofiles==24.1.0
casbin==1.36.3
user-agents==2.2.0
pydantic-validation-decorator==0.1.5
fastmcp==2.5.2
WMI==1.5.1; sys_platform == "win32"
pywin32==310; sys_platform == "win32"
